tools/content_pipeline/requirements.txt
# YouTube 搜索依赖
yt-dlp
youtube-transcript-api
fake-useragent
requests

# 网页搜索依赖
aiohttp
aiofiles
tenacity

# 环境变量管理
python-dotenv
